Move channel information into gem
There were some configuration things still left in the dummy app that
prevent it from being an independent gem

Some possible work left to do:

* Move entire actioncable setup into frenchtoast
* At least separate out identification from main application

Other notes about branch

* Provide a development branch for dummy app so it stops complaining
  so much
