+++ app/channels/french_toast/notifier_channel.rb
@@ -1,0 +1,11 @@
+module FrenchToast
+  class NotifierChannel < ApplicationCable::Channel
+    def subscribed
+      stream_from session_key
+    end
+
+    def unsubscribed
+      stop_all_streams
+    end
+  end
+end
