+++ spec/dummy/Procfile
@@ -1,0 +1,2 @@
+web: bundle exec puma -p $PORT -C ./config/puma.rb
+worker: bundle exec rake jobs:work
