+++ spec/dummy/config/secrets.yml
@@ -1,5 +1,8 @@
 default: &default
   secret_key_base: <%= ENV["SECRET_KEY_BASE"] %>
 
+development:
+  <<: *default
+
 test:
   <<: *default
